options-overhaul: split up options.cpp
